Fixed A Bottleneck caused by empty Playlist check
As Mentioned in #20 The Performance with larger list was very sluggish.
PR #35 tried fixing it with reducing the server request. But during
Empty Playlist checks there were constant server requests, which caused
a bottleneck and with list sizes of greater than 6000 there was a slow
down. This commit tries to fix that by checking the local Playlist.