Fixes #52
Previously gomp increased the cpu usage of one of the cores to 100%. It
was because of the notificationRoutine that continuously checked if the
notificationQueue had entries. Initial workaround that I suggested was
adding time.Ticker to the notificationRoutine. This caused a delay in
displaying notifications.

The Current Implementation fixes this by using a buffered channel that
is of the size of the allowed maximum notifications.